Introduce testing and storage concepts
The interface for the store is a convenience which allows for an
in-memory implementation which is used during unit testing. The remote
store will need to be tested via integration testing. This allows the
business logic and the storage logic to be tested independently, and
only becomes problematic if business logic sneaks into the storage
implementation.