Shift to using kotlinx serialization and add exceptions
Why fight with reflection and json and such when someone has already
done it? There is no realy reason (yet) for me to force the multi-field
aspect of storage. And so just getting the json blob working solves the
storage problem. Once I need to be able to search by other fields I may
change the approach.