Begin implementing DynamoDB and Spring
These two end up related due to the need for dependency management and
IoC. Spring will make the project runnable, DDb gives me a way to do
some testing, and will be a decent datatstore at the end of the day.
This isn't necesarily final, but it is the direction I'm likel to go.